+++ HappyTechSystem/HappyTechSystem/Email.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Prototyping
+namespace HappyTechSystem
 {
     class Email
     {
