+++ HappyTechSystem/HappyTechSystem/Program.cs
@@ -17,6 +17,7 @@ namespace HappyTechSystem
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new Form1());
+            
         }
     }
 }
